Pack all PCI addresses to integers
The PCI address returned by `udev_device_get_sysname()` is represented with
an ASCII string. It is inefficient format for any post-processing and is
also wasteful in memory usage.

Introduce `udev_read_pci_addr()`, which parses the string presentation and
packs it into `uint32_t`.

Further, this unblocks the radix sort also for sorting PCI devices, in
addition to IOMMU groups.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>