json: implement a custom builder
Implement a custom builder for JSON graph in order to drop libcjson dep.

The renderer does not have any visual representation capapbilities i.e., it
simply emits one huge string. It's easy enough to tweak that with widely
available tools such as jq.

In addition, refresh the dependency list in README.md because it is
completely out of sync.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>